+++ src/strategies/agent_strategy_factory.py
@@ -1,4 +1,5 @@
 from .single_agent_rag_strategy_v1 import SingleAgentRAGStrategyV1
+from .single_agent_rag_strategy_v2 import SingleAgentRAGStrategyV2
 from .agent_strategies import AgentStrategies
 from .mcp_strategy import McpStrategy
 from .nl2sql_strategy import NL2SQLStrategy
@@ -9,7 +10,9 @@ class AgentStrategyFactory:
         """
         Return an instance of the strategy class corresponding to the given key.
         """
-        if key in (AgentStrategies.SINGLE_AGENT_RAG.value, AgentStrategies.SINGLE_AGENT_RAG_V1.value):
+        if key == AgentStrategies.SINGLE_AGENT_RAG.value:
+            return await SingleAgentRAGStrategyV2.create()
+        if key == AgentStrategies.SINGLE_AGENT_RAG_V1.value:
             return SingleAgentRAGStrategyV1()
         if key == AgentStrategies.MCP.value:
             return await McpStrategy.create()